pyproject.toml
[tool.poetry]
name = "kane"
version = "0.1.0"
description = ""
authors = ["Takuya Kanesaki"]

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.34.0"
fastapi = "^0.115.8"
python-dotenv = "^1.0.1"
pyjwt = "^2.10.1"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
